pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "doxygen-guard"
version = "0.1.0"
description = "Pre-commit hook that validates doxygen comments for presence, version staleness, and custom tag syntax"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Tyler VanFossen" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pyyaml>=6.0",
]

[project.scripts]
doxygen-guard = "doxygen_guard.main:cli_main"

[project.urls]
Homepage = "https://github.com/tvanfossen/doxygen-guard"
Issues = "https://github.com/tvanfossen/doxygen-guard/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]
